+++ NEWS.md
@@ -11,6 +11,8 @@
 -   Fix errors on `slice_min()` and `slice_max()` for SpatRasters (#20). Also
     add a new parameter `na.rm`.
 
+-   Add new scale: `scale_fill_topo_c()`
+
 # tidyterra 0.1.0
 
 -   Add DOI.